refactor: separate components for auth
