+++ src/routes/(auth)/auth/(components)/index.ts
@@ -1,0 +1,3 @@
+export { default as LoginForm } from './login-form.svelte';
+export { default as RegisterForm } from './register-form.svelte';
+export { default as SSOForm } from './sso-form.svelte';
